+++ packages/grafana-data/src/transformations/matchers/fieldTypeMatcher.test.ts
@@ -1,16 +1,8 @@
-import { toDataFrame } from '../../dataframe/processDataFrame';
 import { FieldType } from '../../types/dataFrame';
 import { fieldMatchers } from '../matchers';
 
 import { FieldMatcherID } from './ids';
-
-export const simpleSeriesWithTypes = toDataFrame({
-  fields: [
-    { name: 'A', type: FieldType.time },
-    { name: 'B', type: FieldType.boolean },
-    { name: 'C', type: FieldType.string },
-  ],
-});
+import { simpleSeriesWithTypes } from './mocks';
 
 describe('Field Type Matcher', () => {
   const matcher = fieldMatchers.get(FieldMatcherID.byType);
