+++ packages/grafana-data/src/transformations/matchers/mocks.ts
@@ -1,0 +1,10 @@
+import { toDataFrame } from '../../dataframe/processDataFrame';
+import { FieldType } from '../../types/dataFrame';
+
+export const simpleSeriesWithTypes = toDataFrame({
+  fields: [
+    { name: 'A', type: FieldType.time },
+    { name: 'B', type: FieldType.boolean },
+    { name: 'C', type: FieldType.string },
+  ],
+});
