+++ packages/grafana-data/src/transformations/matchers/predicates.test.ts
@@ -2,8 +2,8 @@ import { FieldType } from '../../types/dataFrame';
 import { MatcherConfig } from '../../types/transformations';
 import { fieldMatchers } from '../matchers';
 
-import { simpleSeriesWithTypes } from './fieldTypeMatcher.test';
 import { FieldMatcherID, MatcherID } from './ids';
+import { simpleSeriesWithTypes } from './mocks';
 
 const matchesNumberConfig: MatcherConfig = {
   id: FieldMatcherID.byType,
