+++ public/app/core/services/FetchQueue.test.ts
@@ -1,24 +1,9 @@
-import { Observable } from 'rxjs';
 import { take } from 'rxjs/operators';
 
 import { BackendSrvRequest } from '@grafana/runtime';
 
 import { FetchQueue, FetchQueueUpdate, FetchStatus } from './FetchQueue';
-
-type SubscribeTesterArgs<T> = {
-  observable: Observable<T>;
-  expectCallback: (data: T) => void;
-  doneCallback: jest.DoneCallback;
-};
-
-export const subscribeTester = <T>({ observable, expectCallback, doneCallback }: SubscribeTesterArgs<T>) => {
-  observable.subscribe({
-    next: (data) => expectCallback(data),
-    complete: () => {
-      doneCallback();
-    },
-  });
-};
+import { subscribeTester } from './__mocks__/subscribeTester';
 
 describe('FetchQueue', () => {
   describe('add', () => {
