+++ public/app/features/explore/TraceView/components/TracePageHeader/mocks.ts
@@ -1,0 +1,128 @@
+export const trace = {
+  services: [{ name: 'serviceA', numberOfSpans: 1 }],
+  spans: [
+    {
+      traceID: '164afda25df92413',
+      spanID: '264afda25df92413',
+      operationName: 'HTTP Client',
+      serviceName: 'serviceA',
+      subsidiarilyReferencedBy: [],
+      startTime: 1675602037286989,
+      duration: 5685,
+      logs: [],
+      references: [],
+      tags: [],
+      processID: '264afda25df92413',
+      flags: 0,
+      process: {
+        serviceName: 'lb',
+        tags: [],
+      },
+      relativeStartTime: 0,
+      depth: 0,
+      hasChildren: false,
+      childSpanCount: 0,
+      warnings: [],
+      childSpanIds: [],
+    },
+    {
+      traceID: '164afda25df92413',
+      spanID: '364afda25df92413',
+      operationName: 'HTTP Client',
+      serviceName: 'serviceB',
+      subsidiarilyReferencedBy: [],
+      startTime: 1675602037286989,
+      duration: 5685,
+      logs: [],
+      references: [],
+      tags: [
+        {
+          key: 'http.url',
+          type: 'String',
+          value: `/v2/gamma/792edh2w897y2huehd2h89`,
+        },
+        {
+          key: 'http.method',
+          type: 'String',
+          value: `POST`,
+        },
+        {
+          key: 'http.status_code',
+          type: 'String',
+          value: `200`,
+        },
+      ],
+      processID: '364afda25df92413',
+      flags: 0,
+      process: {
+        serviceName: 'lb',
+        tags: [],
+      },
+      relativeStartTime: 0,
+      depth: 0,
+      hasChildren: false,
+      childSpanCount: 0,
+      warnings: [],
+      childSpanIds: [],
+    },
+    {
+      traceID: '164afda25df92413',
+      spanID: '464afda25df92413',
+      operationName: 'HTTP Server',
+      serviceName: 'serviceC',
+      subsidiarilyReferencedBy: [],
+      startTime: 1675602037286989,
+      duration: 5685,
+      logs: [],
+      references: [],
+      tags: [
+        {
+          key: 'http.url',
+          type: 'String',
+          value: `/v2/gamma/792edh2w897y2huehd2h89`,
+        },
+        {
+          key: 'http.method',
+          type: 'String',
+          value: `POST`,
+        },
+        {
+          key: 'http.status_code',
+          type: 'String',
+          value: `200`,
+        },
+      ],
+      processID: '464afda25df92413',
+      flags: 0,
+      process: {
+        serviceName: 'db',
+        tags: [],
+      },
+      relativeStartTime: 0,
+      depth: 0,
+      hasChildren: false,
+      childSpanCount: 0,
+      warnings: [],
+      childSpanIds: [],
+    },
+  ],
+  traceID: '8bb35a31-eb64-512d-aaed-ddd61887bb2b',
+  traceName: 'serviceA: GET',
+  processes: {
+    '264afda25df92413': {
+      serviceName: 'serviceA',
+      tags: [],
+    },
+    '364afda25df92413': {
+      serviceName: 'serviceB',
+      tags: [],
+    },
+    '464afda25df92413': {
+      serviceName: 'serviceC',
+      tags: [],
+    },
+  },
+  duration: 2355515,
+  startTime: 1675605056289000,
+  endTime: 1675605058644515,
+};
