+++ public/app/plugins/datasource/grafana-pyroscope-datasource/datasource.test.ts
@@ -1,37 +1,10 @@
-import {
-  AbstractLabelOperator,
-  CoreApp,
-  DataSourceInstanceSettings,
-  PluginMetaInfo,
-  PluginType,
-  DataSourceJsonData,
-  makeTimeRange,
-} from '@grafana/data';
-import { getBackendSrv, setBackendSrv, TemplateSrv } from '@grafana/runtime';
+import { AbstractLabelOperator, CoreApp, makeTimeRange } from '@grafana/data';
+import { TemplateSrv } from '@grafana/runtime';
 
 import { defaultPyroscopeQueryType } from './dataquery.gen';
 import { normalizeQuery, PyroscopeDataSource } from './datasource';
+import { defaultSettings, mockFetchPyroscopeDatasourceSettings } from './mocks';
 import { Query } from './types';
-
-/** The datasource QueryEditor fetches datasource settings to send to the extension's `configure` method */
-export function mockFetchPyroscopeDatasourceSettings(
-  datasourceSettings?: Partial<DataSourceInstanceSettings<DataSourceJsonData>>
-) {
-  const settings = { ...defaultSettings, ...datasourceSettings };
-  const returnValues: Record<string, unknown> = {
-    [`/api/datasources/uid/${settings.uid}`]: settings,
-  };
-  setBackendSrv({
-    ...getBackendSrv(),
-    get: function <T>(path: string) {
-      const value = returnValues[path];
-      if (value) {
-        return Promise.resolve(value as T);
-      }
-      return Promise.reject({ message: 'reject' });
-    },
-  });
-}
 
 function setupDatasource() {
   mockFetchPyroscopeDatasourceSettings();
@@ -174,22 +147,4 @@ const defaultQuery = (query: Partial<Query>): Query => {
     queryType: defaultPyroscopeQueryType,
     ...query,
   };
-};
-
-const defaultSettings: DataSourceInstanceSettings = {
-  id: 0,
-  uid: 'pyroscope',
-  type: 'profiling',
-  name: 'pyroscope',
-  access: 'proxy',
-  meta: {
-    id: 'pyroscope',
-    name: 'pyroscope',
-    type: PluginType.datasource,
-    info: {} as PluginMetaInfo,
-    module: '',
-    baseUrl: '',
-  },
-  jsonData: {},
-  readOnly: false,
 };
