+++ public/app/plugins/datasource/tempo/SearchTraceQLEditor/TagsInput.test.tsx
@@ -8,7 +8,7 @@ import { initTemplateSrv } from '../test/test_utils';
 import { Scope } from '../types';
 
 import TagsInput from './TagsInput';
-import { v1Tags, v2Tags } from './utils.test';
+import { v1Tags, v2Tags } from './mocks';
 
 describe('TagsInput', () => {
   let user: ReturnType<typeof userEvent.setup>;
