+++ public/app/plugins/datasource/tempo/SearchTraceQLEditor/utils.test.ts
@@ -5,6 +5,7 @@ import { TempoDatasource } from '../datasource';
 import TempoLanguageProvider from '../language_provider';
 import { intrinsics } from '../traceql/traceql';
 
+import { emptyTags, testIntrinsics, v1Tags, v2Tags } from './mocks';
 import {
   filterToQuerySection,
   getAllTags,
@@ -155,21 +156,3 @@ describe('filterToQuerySection returns the correct query section for a filter', 
     expect(result).toBe('span.foo=~"bar|baz"');
   });
 });
-
-export const emptyTags = [];
-export const testIntrinsics = uniq(['duration', 'kind', 'name', 'status'].concat(intrinsics));
-export const v1Tags = ['bar', 'foo'];
-export const v2Tags = [
-  {
-    name: 'resource',
-    tags: ['cluster', 'container'],
-  },
-  {
-    name: 'span',
-    tags: ['db'],
-  },
-  {
-    name: 'intrinsic',
-    tags: testIntrinsics,
-  },
-];
